pyproject.toml
[tool.poetry]
name = "biohackathon-2022"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "biohackathon_2022"}]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
requests = "^2.28.1"
jupyterlab = "^3.5.0"
pandas = "^1.5.1"
matplotlib = "^3.6.2"
seaborn = "^0.12.1"
missingno = "^0.5.1"
lxml = "^4.9.1"
tqdm = "^4.64.1"
nltk = "^3.7"
ipdb = "^0.13.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
